Add support for building Docker image for both x64 and ARM64 platforms (#97)
* Add support for building Docker image for both x64 and ARM64 platforms

Motivation
----------

Currently, this Docker container is only built for x64 architectures. This causes issues on systems that use ARM64 architectures like Apple Silicon.

Docker has had a facility for creating multiplatform images for a while now (via buildx) and the Github docker build action already supports this capability. This PR updates the Github action to specify the platform types to build a multiplatform Docker image.

Modifications
-------------

In both the build and release workflows, I added the 'platform' attribute to the docker build action. The value of the attribute is a comma-separated list of the supported architectures (linux/amd64 for x64 and linux/arm64 for ARM64).

Result
------

The result is the sendgrid mock Docker container will now be built for both x64 and ARM64 platforms.

* Use multiple runners

* Fix runner name

* Remove g++, make, and python from run image.

---------

Co-authored-by: Matt Bertolini <468163+mattbertolini@users.noreply.github.com>
Co-authored-by: Siyavash Habashi <ghashange@live.com>